vis.py: return curve fir parameter for calc_plot_linear_fit
We need to save the linear fit parameters for log plots,
but currently these parameters are just printed on terminal
and not returned by the fir function so that they may be
saved in a csv. This patch adds a return statement returning
the fit parameters.

Issue link: https://github.com/prasadtalasila/IRCLogParser/issues/96