install.sh: download required nltk corpus
The nlkt corpus wordnet is being used and needs
to be installed by the script. It cannot be put
inside setup.py since nltk corpus are not library
that can be installed by pip.
This patch adds a line in install shell script
to download the wordnet corpus.

Issue link: https://github.com/prasadtalasila/IRCLogParser/issues/135