Integrate slackware with dev branch
Since dev branch has undergone lot of changes,
to be able to port the slackware to dev we need
to modify the current dev branch to incorporate
line parsing difference with slack.
This commit modifies dev branch to integrate
slack and also ad some unit-test for further
proofing out analysis.